Use BEST_COMPRESSION (level 9) when deflating patch and file content
The compression speed of the backend server doesn't matter, but the
decompression speed does.  Since zlib inflate() performance is not
tied to the compression level used during deflate() we can reduce
our data size (and thus the size of the data in the data store)
by packing it tighter, without any runtime read penalty.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 0518942342a94a9154da95ccd8966283786b5d0f