Show the unified diff view when the patch cannot be parsed
If the file is really big we may wind up showing a truncated diff
which simply says "File content is too large to display", somewhat
like how a binary file will display a message indicating the files
are different, but binary and thus not suitable for a patch format.

In such cases the patch parser cannot parse the patch, so instead
show the user the unified diff view of the file.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 61b4e782552ef9c274b7ee3bb9b08ce39bb447b7