Fix path conflict merge failures to unsubmit the change
When a path conflict was detected during a merge Gerrit failed to
unsubmit the change due to a data store API usage bug in the merge
service.  We actually marked the change unmerged, wrote it to the
data store, but then used a stale copy of the change (from prior
to the unsubmit) to record the path conflict message.  This meant
we overwrote the merge status flags, so the change still thought
it was pending merge.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 6c38051ae5bbe2282901abc5a4a794da619daca0