Remove pointless is_code_reviewed function from Project, Branch
We always require code reviews.  This is a code review system.

At one point we had considered allowing specific projects (or
branches within a project) to bypass the code review requirements
to make rapid development easier, but we're unlikely to get to
that anytime in the near future.  Most of the approval code
needs a serious refactoring before we try to wedge in this
on/off switch.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 117ef5cc3bb2138593bceb003d3d56ceb9c8324e