Display merge commits using 'diff --cc' rules
A merge commit shouldn't show the total set of changes introduced
by a side branch; its simply too verbose and may tend to break the
100 file limit anyway.

What users really want to see is the "git diff --cc" rules, which
is to show only locations where there was an honest conflict that
had to be fixed by hand.  This way the merge resolution can get
reviewed prior to submission.

Rietveld wasn't really built to show an arbitrary N-way merge like
Git can produce, so we punt right now and only display the unified
diff for a merge where conflicts are detected.  Perhaps in the
future we can display an N-way side-by-side view (where N is an
arbitrary number over 2).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 6793b9e0fc44ee8dc1bddac72ee5d6feeb9820cc