Fix Change.merge_patchset property to use correct name for the key
The correct name for the merge_patchset_key field of the property
is "merge_patchset_key", not "_merge_patchset_key".  A "feature"
of Google App Engine allowed us to use the latter form (with
the leading underscore) without any issue, up until last night's
production delivery.  As a result of the production server upgrade
this alias no longer works.

It was a bug in Gerrit that we were even using "_X" to reference
a StringProperty named "X".  I made a typo in the code.  But none
of our testing showed the typo because it always worked.  Right up
until the GAE servers received an upgrade and started to behave
differently.

No matter how much you test Python code, there really is no
substitute for a static compiler that can at least verify all
symbolic references being made by the code.  Any decent static
compiler would have identified the typo and forced me to correct it
when I made it.  Python allowed me to make it, and GAE allowed the
typo to work due to an implementation detail of the data store API.
That detail changed, and we got caught.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 31e77e4d3918aa4b0cea05879ac09dd442f813ed