Force the manager application umask to 0022
In general we want all files we create to be world-readable, as the
manager keeps track of open source source code repositories which
are freely distributed.

Recent changes to some of the enviroments where mgrapp runs have
modified the default umask, making me realize we didn't specify
one for the mgrapp itself.  To avoid environment specific errors
we should list exactly what we expect.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 7d644cddd1e83eb204b3b1f5c76b1517418ef385