Implement OpenID based single-sign-on for Google Accounts
This is the early part of our OpenID based SSO implementation.
We now support authentication through the Google Accounts OpenID
provider, allowing users to "sign in" and "sign out" of the
application through the external provider.

Presently the login servlet is hard-coded to only use Google Accounts
as an OpenID provider.  We should support other providers by setting
up the standard OpenID entry box as the initial content returned
from our login dialog.

The login servlet randomly generates its coookie encryption key
during startup.  This causes any existing user cookies to become
invalid, requiring the user to authenticate again.  At some point
this needs to change to obtain the encryption key from a local
store, where it isn't part of the code base but is preserved across
application server restarts.

In the future we should also support other web based login systems,
such as any in-house web application (such as those used by many
large companies) or just plug in a basic LDAP or database based web
interface, without requiring the complexity of an OpenID provider.

The left and top menu now updates to reflect the current status of
the user.  If the user signed in we show the "mine" links and we
offer the admin menu and a link to the user's account settings.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 49a7115213a7889705c7f345c25e7af4f654c502