Set the path of our cookies to only our servlet context
This way if we are hosted on the same web server as other applications
we don't leak our account cookies into their requests.  Its unlikely
that another application on the same server is "evil" and would try
to abuse the cookie if presented, but there is little point in taking
a risk with that.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: cfc2569c374c96116d9ab921c89f947b8cebbafd