Start the account sequence at 1,000,000 so it is different from groups
When looking at an account and/or group id we know anything over the
initial value of 1,000,000 means it is a user account.  We are quite
unlikely to see 1 million groups created within a single Gerrit, so
pushing the user accounts up above this level makes them roughly not
in the same numeric space.  This should make debugging easier.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 9f0c90148d7393705cfab891c0fd031bdae714e3