Allow publishing comments with non-ASCII characters in a user's name
We need to force the strings to be unicode so we can put them into
the data store.  I don't know why this is necessary.  String handling
in Python is a mess, we get data from the data store from the web
that should be a unicode string but half the time it loses its
unicode status.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 156c1fcca11d54ad9ba6c888ac5685a536873cf8