Add ChangeApproval and flexible ApprovalCategory definitions
Some workflows based on Gerrit may need one or more ways to let a
user stamp a change with some sort of "OK" or "not-OK" indication.
The basic Gerrit 1 workflow had two such stamps, of "Verified" and
"Code Review".

Users however have asked to add or remove these steps in their
workflow, so we now have a flexible ApprovalCategory definition
which defines a range of permissible numeric values.  For any
given change each Account is permitted to provide a value for
each ApprovalCategory, or to omit it entirely.

The schema isn't normalized the way I'd like it to be.  We have
to use a unique entity for each "cell" in the review status table
(the user + category combination) because we cannot generate our
schema on the fly.  Ideally we would have created a unique column for
each category in the database table, and held only 1 row per user.
Since we only ever care about the entire table for a change we can
still at least fetch the entire cell array in one query, but we
will get back more individual entities.  Its slightly higher data
store API overhead in exchange for a bit of needed flexibility,
without hacking up gwtorm.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 61056e5be8e71d8c72ff45468ee8af3bee541a89