Add DeltaContent entity to store compressed file contents
This is exactly like Gerrit 1, where the content is a unified
diff that can be applied on the fly to reconstruct the full file.
It worked well for us in Gerrit 1, by giving very tight compression
but reasonably efficient access, so long as we bound the delta
chain depth at a reasonable figure (like 10).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 7b6ecc91eb5b2107b62214a81c72c41e0f93c0d3