export_safe: A script to scrub and export a Gerrit2 database
This script is just a simple tool to help me export the current
PostgreSQL database into a format that is safe to send to a 3rd
party.  All of the data is public on Gerrit's website, except for
the user's physical address contact information.  We scrub that
out because we do not have consent from our users to disclose
that information.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2f09570d608976d34d9ae0ffa892e6ce00c6d1af