Move Screen to our new client.ui and modify its title to use <h1> tags
Now that we have client.ui for basic UI widgets that are common to
all of Gerrit we can move our Screen class into it.

The title for a screen is now drawn in an <h1> element, making it
much more accessible to screen readers.  Using a GWT Label is sort
of heavy-weight for such a simple concept as our title widget.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 53046651df0f91abae05ac67c474191e03765ea2