Delay invocation of SignedInListener.onSignIn to fix Firefox bug
Calling the SignedInListeners here is way to early. What happens is the
call stack is based on the "jsonp" style callback being made inside of
the dialog's iframe.  Trigging RPCs from within that callstack means
we are unable to read the RPC results, which sounds pretty odd.  But if
we delay the listeners (which may want to make RPC calls) until later
then things are just fine.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 88f26588b1667e3b017e26a6e0b7334576dd91f1