Load the user's Account data at module load so we can show their name
We now display the user's name in the menu bar, like Gerrit 1 did.  This
is based on an RPC call to load the user's account information into the
module as a static member; if the cookie is valid we will get their data
back.  If it has expired or is forged, we'll get no account and give them
an anonymous account page instead.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 0c8fb8c99e3d1a115ce1c1ef3247156c99933f4d