Define a filter to force GWT compiler output to be cached indefinitely
GWT output can be cached essentially forever, as any change to the
content will cause a change in the URL, resulting in a different
resource to be loaded, and a cache miss.  Setting the cache time
to a full year gives plenty of flexibility to edge caches to avoid
even revalidating their cache entries before serving them.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: bdb917d2f8dff89ad7b13d1a2e89d261ec827919