Ensure logins happen via the application's canonical URL
Google's OpenID provider issues a different token for each hostname/port
combination it receives from the relying party.  If Gerrit is published
under a CNAME like "r.android.com" we need to make sure the URL that we
use to refer to Google by is always the same, no matter what hostname
the user tried to reach us through.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 99aebba40065c3e12fb6b3f23e609ec71d7b88eb