Catch RPC exceptions and display them in a modal dialog
Currently we really can't recover automatically from any RPC error,
so instead of failing silently in compiled mode (due the compiler
stripping out GWT.log calls) we show a modal popup to the user to
let them know things didn't go according to plan.

In the case of a star toggle, we untoggle it, but pretty much all
of the other actions don't really have a reverse, so we use the
default error display logic in the dialog.

Future versions of gwtjsonrpc should support automatically retrying
certain types of requests, especially when we can identify the error
as a network connectivity problem.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 44b72d261ccea965d060417247174eb1306f16c4