Send global system configuration data to the client on module load
We need some key data about how this Gerrit has been setup on the
client side; in particular the approval categories and their set
of known values will be needed frequently in the different change
display screens.  Caching this in a static within the page will
reduce the size of each individual RPC, as we only need to send
this information once.

The data is usually shoved into the top of the host page, so it is
available immediately when the module initializes itself.  If it
isn't there an RPC is used to load the data, but this should only
happen while debugging under hosted mode.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 304ef2dcf5ea630bc799cc4f47d1cefeeebfc804