Cleanup formatting on Firefox 2, 3 and Safari
The sign in dialog wasn't quite tall enough, now it is.  But its
far too large for the Android web browser.

The change info block is now folded up with the description, fixing
the layout problems it was having on Firefox 3 where it forced the
description down lower on the page, and on Firefox 2 when the window
was too narrow.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 59ee46c677c89a90c8e315d2264a990f504d8886