Import all patch data again from Git to ensure it is accurate
We've made some schema changes about how patch data is stored and
represented in the Gerrit 2 database, so we need to import all of
the patch entities over again.  This change includes a CLI tool
that can be run to re-import one or more patch set entities from
the corresponding Git commit.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 38270b642ddb2c457104f07de2485d999ea60ae2