Change our no-cache headers to match gwtjsonrpc
gwtjsonrpc now uses a fixed expire date back in the past, so its
quite clear that the entity shouldn't be cached by a proxy or
a browser.  Its also faster, as we don't have to format the date.

We also now mark must-revalidate on our Cache-Control, so a proxy
server doesn't try to serve stale content while waiting for a reply
from Gerrit.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 0bf5c68e9c1c809be0079525b45319f372b408c5