Allow users to link additional web identities to their Gerrit account
This is really useful with Google Accounts, where the user might have
more than one Google Account identity in their browser.  If they link
all of them back to the same Gerrit account then it doesn't matter who
they are currently logged in as in their browser, they can still get
to the same dashboard and the same identity within Gerrit, without a
forced logout/login cycle on the google.com domain.

We store the email address uniquely with each identity so the user
can more easily tell them apart.  In the case of Google Accounts the
identity string is too opaque for an end-user to deduce which account
is which without having the email along side of it.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: bd5d5ef82316866a19e910844ea116448d893bc1