Redo precompression of *.html,*.css when creating the WAR file
Back in 688dd17e965 [formerly 1176f4538d4] ("Build a standard .war file for Gerrit") we
precompressed the HTML and CSS files so Jetty can server them
as-is to the client, reducing the network and CPU required for
the very large GWT generated JavaScript.  When we switched over
the build to using Maven we lost that precompression logic.

This change adds it back in, depending upon the "gzip" command
to be in the caller's $PATH.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 56d76c11a7dc45543ed4da850fc44fea74d8ce77