Fix the patch import count to match what we actually imported
Gerrit1's data store contains orphaned patch records, where the
patch set and the change have been removed from the data store
but the patch entity is still there.  Because we don't process
those into Gerrit 2 our patch counts were always lower than in
Gerrit 1.  We should account for that when reporting the summary
at the end of the import.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: c2b3192aecf9ef126da94a7db635de617b81dc60