Refactor GerritCommandFactory to lookup by hash
Using a HashMap to locate the commands is likely quicker than doing
a linear scan through the different support names.  Its also a lot
easier to read when we insert a new command implementation into the
table, as we aren't adding an if test, its just a new registration
into the map.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 4f185243373c9ed4da71e3ec8fa901f4ec1c1f22