Import approval right entities from Gerrit 1 into ProjectRight
The ProjectRight entity binds a single group to an ApprovalCategory,
and denotes what range of values within the ApprovalCategory can be
used by the members of that group.

Since Gerrit 1 supported individual users to be added to projects in
the approver, verifier or submitter category we create fake groups for
these cases and toss the other users into those groups.

The submitter category uses a position of -1, indicating that it should
not appear in the review status table of a change, but it still is a
valid category within the system.  This hack is necesary to reuse the
same ProjectRight system to handle the submitter status.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: a26dc96cf87f98286ee8b51bbe0820e8b67343cb