Require the committer email address to match one on file
The committer field must use an email address part which appears
in one of the external identities for the logged in account.  We
verify this to ensure that changes are attribute correctly back
to the original submitter.

Only committer is validated; author is permitted to be free-form
so that patch based workflows where "git am" set the author field
from an email message's From header still work correctly.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 02f83862efe22c3c26f93e2cb0d28315b8e81918