Clear any ChangeApprovals already placed when adding a patch set
This way an ACK on patch set 1 doesn't allow a submitter to
upload a pack set 2, bypass the ACK status, and jump right
to the submission queue.

We only clear positive ACKs under the assumption that a
negative one has good reason to stand, and may need to be
resolved manually.  Clearing positive ACKs just denies the
"replace and submit" hole in the workflow.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: e0ffe1f6f643669dea8708d097c1a353f71d8e62