Automatically set each approval category for the current user
During receive we assume the submitter wants to approve this change,
so we automatically insert the ChangeApproval records for all known
workflow stops at their max value.  The user might not have access
to these, but that will be fixed later at display time in case the
ProjectRights are updated post-upload.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 3937647979e58b03e474dda9b0137b0a95bfecd8