Normalize the change approvals on unclosed changes
If a change hasn't been closed yet its approvals are still in flux.
We normalize the value prior to display based on the current values
in the ProjectRights entities for the destination project.

Any approval category not at its max is listed in the new member of
ChangeDetail so clients can determine what status is missing from
this change.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2e2d5a858279c4e77d5a65ed04f038dcb4e0bbb6