Move the SchemaFactory<ReviewDb> into Common
Some day when we have a pure JS offline mode client we'll need
access to our schema factory for the offline versions of our
service implementations.  Moving it to Common as a static field
cleans up a lot of our code, as we don't have to pass around the
GerritServer instance quite as much.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: ab79bcb54b25d87bf9fa4d4ca765dfacfe15f75b