Remove ContributorAgreement.groupAccess field as its unused
I probably need to rethink the group agreement code.  Really what
we want is to attach an agreement to a group within Gerrit, then
give those who have signing authority on that group agreement the
ability to manage that group's membership.  Anyone currently in
that group is considered to have signed the group agreement.  Its
certainly not structure that way right now though.

Since this field is unused I'm dropping it out of the schema to
avoid any confusion.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 8679613fc0fab89c569deec548bd4a563fc10d2f