Replace RightRule with a more flexible (and cleaner) implementation
The CategoryFunction implementations control how they compute their
valid status, and have access to fast caches of all information for
the users involved and the project's granted rights.

Two implementations of CategoryFunction are provided, giving us the
classical features we had in Gerrit1 for the verified, code review
and submit functionality.

In the future we may add more functions to the stock system, or
provide a way for these to be scripted in some lightweight scripting
language like JavaScript, so the function can be executed on the
browser and on the server (by Mozilla Rhino).

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 9a6aca9cdc8d30586c21ab671e9ad7083b6655ff