Change export_safe.sh to dump the gerrit1 schema
The gerrit1 schema doesn't change, so its easier to dump that and
run the conversion than to keep dumping the gerrit2 schema anytime
I need to send a schema dump to another server.

I also updated the conversion instructions to match what I just
executed; apparently it was a bit out of date relative to the
process I have been using.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 2240f799b3b3c34eaf3d4aa0cb5e3fba72669813