Linkify bare hyperlinks in messages
Any hyperlinks in commit messages or in comments on a a patch file
or a change are now setup inside of an <a href> tag so users can
actually click on these in their browser.  It slows down rendering
a tad as we have to slurp through the text with a regex, but its
well worth it when there actually is a link present.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: 4aeceacd7a668b9168ebfde9ab884397588b450c