Switch backup to use pickles over RPC instead of XML
The to_xml() method of a datastore entity updates the auto_now
date time fields.  This is a known bug:

  http://code.google.com/p/googleappengine/issues/detail?id=322

Since we cannot fix it we have to serialize everything in pickles
and then use the pickled format on the client side, which means we
also must import the GAE SDK.  I've hardcoded it to where it is on
my system, which is ../../google_appengine, aka a sibling of the
gerrit repo client.

Signed-off-by: Shawn O. Pearce <sop@google.com>

Former-commit-id: eed6f5a863fb6a28ff6c380c91a49c9043b40e5a